backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
motor==3.7.1
python-dotenv==1.0.0
pydantic-settings==2.0.3
email-validator==2.0.0
PyJWT==2.10.1
redis==5.2.1
prometheus-client==0.21.1
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.27.2
